[BACKEND-13] refactor(errors): move repository errors into db package
